Updated script based on Debian testing

+++ scripts/virtualbox.sh
@@ -6,6 +6,7 @@ set -x
 codename="$(facter lsbdistcodename)"
 os="$(facter operatingsystem)"
 os_family="$(facter osfamily)"
+os_release="$(facter operatingsystemrelease)"
 
 if [ "$PACKER_BUILDER_TYPE" != "virtualbox-iso" ]; then
     exit 0
@@ -17,12 +18,20 @@ if [[ $os_family == "Debian" ]]; then
         sudo rm -rf /home/vagrant/VBoxGuestAdditions*.iso
         
         elif [[ $os == "Debian" ]]; then
-        sudo mkdir -p /mnt/virtualbox
-        sudo mount -o loop /home/vagrant/VBoxGuestAdditions*.iso /mnt/virtualbox
-        sudo sh /mnt/virtualbox/VBoxLinuxAdditions.run
-        sudo umount /mnt/virtualbox
-        sudo rm -rf /home/vagrant/VBoxGuestAdditions*.iso
+        if [[ $os_release > 7.11 ]]; then
+            sudo mkdir -p /mnt/virtualbox
+            sudo mount -o loop /home/vagrant/VBoxGuestAdditions*.iso /mnt/virtualbox
+            sudo sh /mnt/virtualbox/VBoxLinuxAdditions.run
+            sudo umount /mnt/virtualbox
+            sudo rm -rf /home/vagrant/VBoxGuestAdditions*.iso
+        else
+            if [ -f /home/vagrant/VBoxGuestAdditions*.iso ]; then
+                sudo rm -rf /home/vagrant/VBoxGuestAdditions*.iso
+            fi
+            
+        fi
     fi
+    
     elif [[ $os_family == "RedHat" ]]; then
     if [[ $os != "Fedora" ]]; then
         sudo yum -y install gcc kernel-devel kernel-headers dkms make bzip2 perl && \
